requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
openai==1.12.0
protobuf==3.20.3
requests==2.31.0 